Moved placeShipsOnBoard to gameBoard
broke a test for placeShipsOnBoard and i realize
it just doesnt make sense to have a function like
this when place ship exists and i am not sure
how the final function would end up looking.

So i just gonna skip the test and rewrite this
entire function once the game logic is sorted.